// src/compass/eligibility-engine/types.ts
export enum EmploymentStatus {
  employed = "employed",
  selfEmployed = "selfEmployed",
  unemployed = "unemployed",
  student = "student",
  retired = "retired",
  disabled = "disabled",
  notInLaborForce = "notInLaborForce",
}

export enum CitizenshipStatus {
  citizen = "citizen",
  permanentResident = "permanentResident",
  refugee = "refugee",
  asylumSeeker = "asylumSeeker",
  temporaryVisa = "temporaryVisa",
  undocumented = "undocumented",
  other = "other",
}